MS_example/requirements.txt
flask
Flask-SQLAlchemy
Flask-Migrate
requests
psycopg2-binary
numpy
pytest
pandas